% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm625-package.R
\docType{package}
\name{lm625-package}
\alias{lm625}
\alias{lm625-package}
\title{lm625: Fast Linear Regression for Biostat 625 HW4}
\description{
Simple linear regression implemented from scratch for Biostat 625 homework 4.
}
\author{
\strong{Maintainer}: Junyi Zhang \email{zjuny@umich.edu}

}
\keyword{internal}
